--- burgerqueen/Slice/allDay.js
@@ -1,44 +1,0 @@
-const waiterAllDay = () => {`
-    <table>
-        <tr>
-            <td>Price $</td>
-            <td>Product</td>
-            <td>Number</td>
-        </tr>
-        <tr>
-            <td>10</td>
-            <td>Hamburguesa simple</td>
-            <td></td>
-        </tr>
-        <tr>
-            <td>15</td>
-            <td>Hamburguesa doble</td>
-            <td></td>
-        </tr>
-        <tr>
-            <td>5</td>
-            <td>Papas fritas</td>
-            <td></td>
-        </tr> 
-        <tr>
-            <td>5</td>
-            <td>Agua 500ml</td>
-            <td></td>
-        </tr>     
-        <tr>
-            <td>7</td>
-            <td>Agua 750ml</td>
-            <td></td>
-        </tr>     
-        <tr>
-            <td>7</td>
-            <td>Bebida/gaseosa 500ml</td>
-            <td></td>
-        </tr>  
-        <tr>
-            <td>10</td>
-            <td>Bebida/gaseosa 750ml</td>
-            <td></td>
-        </tr>        
-    </table>
-`};
