--- burgerqueen/Slice/menu.js
@@ -1,40 +1,0 @@
-// import React from "react";
-// import logo from '../assets/logo-burgerqueen.png';
-// import Link from "react-router-dom";
-
-// const Menu = () => {
-//     return <div className="div-button-menu">
-//         <div id="logo"><img src={logo} className="App-logo" alt="logo" /></div>
-//         <div className="button-waiter">
-//             <button className="button-orders active">
-//                 <Link to="/">New Order</Link>
-//             </button>
-//             <button className="button-orders">
-//                 <Link to="/pendingOrder">Pending Order</Link>
-//             </button>
-//             <button className="button-orders">
-//                 <Link to="/doneOrder">Done Order</Link>
-//             </button>
-//         </div>
-//     </div>
-// }
-
-// export default Menu;
-
-const menu = () => {`
-    <form>
-        <div>
-            <input type='name'></input>
-        </div>
-        <div>
-            <input type='checkbox'>
-            <label>Morning</label>
-            <label>All day </label>
-            </input>
-        </div>
-        <div id=order></div>
-        <input></input>
-        <button>SUBMIT</button>
-    </form>
-`}
-
