--- burgerqueen/Slice/morning.js
@@ -1,29 +1,0 @@
-const morning = () => {`
-    <table>
-        <tr>
-            <td>Price $</td>
-            <td>Product</td>
-            <td>Number</td>
-        </tr>
-        <tr>
-            <td>5</td>
-            <td>Café americano</td>
-            <td></td>
-        </tr>
-        <tr>
-            <td>7</td>
-            <td>Café con leche</td>
-            <td></td>
-        </tr>
-        <tr>
-            <td>10</td>
-            <td>Sandwich de jamón y queso</td>
-            <td></td>
-        </tr> 
-        <tr>
-            <td>7</td>
-            <td>Jugo de frutas natural</td>
-            <td></td>
-        </tr>                
-    </table>
-`};
