--- burgerqueen/Slice/waiter.js
@@ -1,15 +1,0 @@
-const waiter = () => {`
-    <div>
-        <div>
-            <img src=HAMBURGUESA />
-            <h2>Burger Queen</h2>
-        </div>
-        <div>
-            <button>New order</button>
-            <button>Order pending</button>
-            <button>Done Order</button>
-        </div>
-        <div id='rootWaiter'></div>
-
-    </div>
-`};
