+++ burgerqueen/src/App.css
@@ -8,9 +8,6 @@ A:link {
 }
 
 body {
-  /* height: 100%;
-  width: 100%; */
-  
   background: black;
   font-family: 'Montserrat', sans-serif;
 }
@@ -18,27 +15,13 @@ body {
 /* WAITER LOGIN */
 
 .login-container {
-  /* width: 300px;
-  height: 500px; */
-
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
-
-  
-  /* display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: -10%; */
 }
 
 .logo-welcome img {
-  /* display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center; */
-
   margin-bottom: -30px;
   z-index: -5;
 }
@@ -49,20 +32,11 @@ body {
   text-align: center;
 
   padding-bottom: 1em;
-  /* padding-top: -100px; */
 
   color: yellow;
   font-weight: 600;
   font-size: 20px;
 }
-
-/* .burgerqueen-logo {
-  margin-bottom: -20px;
-} */
-
-/* .form-container {
-
-} */
 
 .formLogin {
   display: flex;
@@ -87,7 +61,13 @@ body {
   width: 170px;
   height: 35px;
   margin-top: .7em;
+
+  color: yellow;
 }
+
+/* .inputLogin ::placeholder {
+  color: yellow;
+} */
 
 .checkboxLogin {
   background: black;
@@ -97,24 +77,7 @@ body {
   margin-top: .7em;
   
   color: yellow;
-  /* background: black;
-  color: yellow;
-  border: 3px solid yellow;
-  width: 170px;
-  height: 35px;
-  margin-bottom: .15em;
-  text-align: center; */
 }
-/* 
-.checkboxLogin option {
-  text-align: center;
-}
-
-.text-checkboxLogin {
- text-align: center;
-} */
-
-
 
 /* WAITER ORDERS */
 
@@ -220,7 +183,7 @@ p {
   font-weight: 400;
 }
 
-.button-send {
+.button-send-waiter {
   border: 3px solid yellow;
   background: none;
 
@@ -234,19 +197,19 @@ p {
   font-size: 15px;
 }
 
-.button-send a {
+.button-send-waiter a {
   color: yellow;
 }
 
-.button-send:hover {
+.button-send-waiter:hover {
   background: yellow;
 }
 
-.button-send:hover a {
+.button-send-waiter:hover a {
   color: black;
 }
 
-.box-send {
+.box-send-waiter {
   display: flex;
   flex-direction: row;
   justify-content: center;
@@ -271,7 +234,7 @@ p {
   justify-content: space-evenly;
 }
 
-.box-one-pending {
+/* .box-one-pending {
   margin-top: 3em;
 
   width: 370px;
@@ -283,24 +246,22 @@ p {
 .box-one-pending p {
   margin-top: 1.5em;
   color: black;
-}
+} */
 
-.box-one-pending, .box-two-pending,
+/* .box-one-pending, .box-two-pending,
 .box-three-pending, .box-four-pending  h3, h4 {
+  text-align: center;
+} */
+
+.box-pending-order h3, h4 {
   text-align: center;
 }
 
 .done {
-  font-size: 1.4em;
+  font-size: 17px;
   font-weight: 700;
 }
-
 /* 
-.box-one-pending, .box-two-pending,
-.box-three-pending, .box-four-pending h4 {
-  text-align: center;
-} */
-
 .box-two-pending {
   margin-top: 3em;
 
@@ -322,9 +283,9 @@ p {
   justify-content: space-evenly;
 
   margin-top: -1.5em;
-}
+} */
 
-.box-three-pending {
+.box-pending-order {
   margin-top: 3em;
 
   width: 370px;
@@ -333,12 +294,12 @@ p {
   background: yellow;
 }
 
-.box-three-pending p {
+.box-pending-order p {
   margin-top: 2em;
   color: black;
 }
 
-.box-four-pending {
+/* .box-four-pending {
   margin-top: 3em;
 
   width: 370px;
@@ -350,7 +311,7 @@ p {
 .box-four-pending p {
   margin-top: 2em;
   color: black;
-}
+} */
 
 .margin-bottom {
   margin-top: 3em;
