+++ burgerqueen/src/components/DoneOrder.js
@@ -1,0 +1,10 @@
+import React from "react";
+import Order from "./Order"
+// import logo from '../assets/logo-burgerqueen.png';
+
+function DoneOrder() {
+    return <Order />
+  }
+  
+export default DoneOrder;
+
