+++ burgerqueen/src/components/Menu.js
@@ -1,0 +1,29 @@
+// import React from "react";
+// import {
+//   BrowserRouter as Router,
+//   Link,
+//   useRouteMatch,
+// } from "react-router-dom";
+// import logo from '../assets/logo-burgerqueen.png';
+
+// const Menu = () => {
+//     const match = useRouteMatch();
+//     return <>
+//     <div className="logo-waiter">
+//         <img src={logo} alt="logo" />
+//       </div>
+//       <div className="container-buttons-order">
+//             <button className="button-order">
+//               <Link to={`${match.url}/`}>New Order</Link> 
+//             </button>
+//             <button className="button-order">
+//               <Link to={`${match.url}/pendingOrder`}>Pending Order</Link> 
+//             </button>
+//             <button className="button-order">
+//               <Link to={`${match.url}/doneOrder`}>Done Order</Link> 
+//             </button>
+//       </div>
+//     </>
+// }
+
+// export default Menu;
