+++ burgerqueen/src/components/PendingOrder.js
@@ -1,135 +1,32 @@
 import React from "react";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link
-} from "react-router-dom";
-import logo from './assets/logo-burgerqueen.png';
-import './App.css';
+// import Order from "./order";
+import Order from "./Order"
 
-export default function App() {
-  return (
-    <Router>
-      <div className="div-button-menu">
-      <div id="logo"><img src={logo} className="App-logo" alt="logo" /></div>
-          <div className="button-waiter">
-                <button className="button-orders active">
-                  <Link to="/">New Order</Link>
-                </button>
-                <button className="button-orders">
-                  <Link to="/pendingOrder">Pending Order</Link>
-                </button>
-                <button className="button-orders">
-                  <Link to="/doneOrder">Done Order</Link>
-                </button>
-         </div>
-         <p>CLIENT_045</p>
-         <div className="main-box">
-            <div className="box-one">
-              <p>Morning</p>
-              <p>All day</p>
-            </div>
-            <div className="box-two">
-              <p>$5  Café americano</p>
-              <p>$7  Café con leche</p>
-              <p>$10 Sándwich jamón y queso</p>
-              <p>$7 Jugo de frutas</p>
-            </div>
-         </div>
-         <p className="total">Total: $</p>
-         <div className="box-send">
-            <button className="button-send">
-                      <Link to="/doneOrder">ENVIAR</Link>
-            </button>
-         </div>
-        {/* A <Switch> looks through its children <Route>s and
-            renders the first one that matches the current URL. */}
-        <Switch>
-          <Route path="/pendingOrder">
-            <About />
-          </Route>
-          <Route path="/doneOrder">
-            <Users />
-          </Route>
-          <Route path="/">
-            <Home />
-          </Route>
-        </Switch>
-      </div>
-    </Router>
-  );
-
-  
-
+function PendingOrder() {
+  return <Order />
 }
 
-function Home() {
-  return <h2>Home</h2>;
-}
-
-function About() {
-  return <h2>About</h2>;
-}
-
-function Users() {
-  return <h2>Users</h2>;
-}
-
-// function pendingOrder() {
-//   let match = useRouteMatch();
-//   return (
-//   <div>
-//     <h2>Topics</h2>
-//     <ul>
-//       <li>
-//         <link to=
-//         {`${match.url}/components`}>Components</link>
-//       </li>
-//       <li>
-//         <link to={`${match.url}/props-v-state`}>
-//           Props v. State
-//         </link>
-//       </li>
-//     </ul>
-//   </div>)
-// }
+export default PendingOrder;
 
 
+// import React from "react";
 
-//import logo from './logo.svg'; 
-// import React, { useState } from "react";
-// import styled from "styled-components";
-// import logo from './assets/logo-burguerqueen.png';
-// import './App.css';
-
-// function App() {
-//   return (
-//     <div className="App">
-//       <header className="App-header">
-//         <img src={logo} className="App-logo" alt="logo" />
-//         <div id="ButtonWaiter">
-//           <button>New Order</button>
-//           <button>Pending Order</button>
-//           <button>Done Order</button>
-//         </div>
-//         <p>
-//           Edit <code>src/App.js</code> and save to reload.
-//         </p>
-//         <a
-//           className="App-link"
-//           href="https://reactjs.org"
-//           target="_blank"
-//           rel="noopener noreferrer"
-//         >
-//           Learn React
-//         </a>
-//       </header>
+// function PendingOrder() {
+//   return <div>
+//     {/* TENER SOLO 1 COMPONENTE QUE NOS MUESTRE LAS TARJETAS DE LAS ORDENES Y QUE VAYA CAMBIENDO CON LOS PROPS O PROPIEDADES QUE TIENE EL COMPONENTE  */}
+//     <div className="main-box-pending">
+//       <div className="box-pending-order">
+//         <h3>Morning</h3>
+//         <h4>CLIENT-048</h4>
+//         <p>$5  Café americano</p>
+//         <p>$7  Café con leche</p>
+//         <p>$10 Sándwich jamón y queso</p>
+//         <p>$7 Jugo de frutas</p>
+//       </div>
 //     </div>
-//   );
+//     <div className="margin-bottom">
+//     </div>
+//   </div>
 // }
 
-// export default App;
-
-
-/*<img src={'./assets/logo-burguerqueen.png'}/>*/
+// export default PendingOrder;
